Got tests to work for random model. Added two basic tests
