+++ src/models/random_model.py
@@ -3,14 +3,14 @@
 """
 import random
 
-from base_model import BaseModel
+from .base_model import BaseModel
 
 class RandomModel(BaseModel):
     def __init__(self):
         self.classes = []
     
     def train(self):
-        pass
+        return "No training needed for this model"
 
     def predict(self):
         return random.choice(self.classes)
