--- tests/models/random_model.py
@@ -1,17 +1,0 @@
-"""
-    This file defines the tests for the random prediction model.
-"""
-import unittest
-
-from src.models import random_model
-
-class RandomModelTester(unittest.TestCase):
-    def test_train(self):
-        pass
-    
-    def test_predict(self):
-        pass
-
-
-if __name__ == "__main__":
-    unittest.main()
